Revert "Ported: Bluetooth Message Access Profile (MAP) from CM9"
This reverts commit 0e5fdcffdd1c3f1292a7cfa04660c966f70c883a.

MAP: Support MAP Server role on Bluedroid with javax.OBEX lib.

Handle No END OF BODY for MAP 1.0 Server Get request. Only the
last packet need send the 0x49(End of Body). For intermediate
packets, we need to send 0x48 (Body).

Move definitions for MAP or FTP Service Specific UUID and
Request type for CONNECTION_ACCESS_REQUEST intent to
BluetoothExt APK.

CRs-fixed: 504042

+++ obex/javax/obex/ServerOperation.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2010-2011, Motorola, Inc.
+ * Copyright (c) 2013, The Linux Foundation. All rights reserved.
+ * Copyright (c) 2008-2009, Motorola, Inc.
  *
  * All rights reserved.
  *
@@ -89,7 +90,6 @@ public final class ServerOperation implements Operation, BaseStream {
     private boolean mHasBody;
 
     private boolean mEndofBody = true;
-
     /**
      * Creates new ServerOperation
      * @param p the parent that created this object
@@ -367,7 +367,7 @@ public final class ServerOperation implements Operation, BaseStream {
                  */
                 if ((finalBitSet) || (mPrivateOutput.isClosed())) {
                     if (mEndofBody) {
-                        out.write((byte)0x49);
+                        out.write(0x49);
                         bodyLength += 3;
                         out.write((byte)(bodyLength >> 8));
                         out.write((byte)bodyLength);
@@ -380,21 +380,19 @@ public final class ServerOperation implements Operation, BaseStream {
                     out.write((byte)bodyLength);
                     out.write(body);
                 }
-
             }
         }
 
-        if ((finalBitSet) && (type == ResponseCodes.OBEX_HTTP_OK) && (orginalBodyLength <= 0)) {
-            if (mEndofBody) {
-               out.write(0x49);
-               orginalBodyLength = 3;
-               out.write((byte)(orginalBodyLength >> 8));
-               out.write((byte)orginalBodyLength);
-           }
-        }
-
-        mResponseSize = 3;
-        mParent.sendResponse(type, out.toByteArray());
+            if ((finalBitSet) && (type == ResponseCodes.OBEX_HTTP_OK) && (orginalBodyLength <= 0)) {
+                if (mEndofBody) {
+                    out.write(0x49);
+                    orginalBodyLength = 3;
+                    out.write((byte)(orginalBodyLength >> 8));
+                    out.write((byte)orginalBodyLength);
+                }
+            }
+            mResponseSize = 3;
+            mParent.sendResponse(type, out.toByteArray());
 
         if (type == ResponseCodes.OBEX_HTTP_CONTINUE) {
             int headerID = mInput.read();
@@ -724,5 +722,4 @@ public final class ServerOperation implements Operation, BaseStream {
     public void noEndofBody() {
         mEndofBody = false;
     }
-
 }
